pyproject.toml
[tool.poetry]
name = "rag"
version = "0.1.0"
description = ""
authors = ["Gilvan Wanderley <gilvan.wanderley@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.34.0"
pypdf2 = "^3.0.1"
langchain = "^0.2.0"
python-dotenv = "^1.0.1"
faiss-cpu = "^1.8.0"
openai = "^1.30.1"
huggingface-hub = "^0.23.1"
langchain-community = "^0.2.0"
langchain-openai = "^0.1.7"
instructorembedding = "^1.0.1"
torch = "^2.3.0"
sentence-transformers = "^2.7.0"
tiktoken = "^0.7.0"
langchainhub = "^0.1.15"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
